# [Perf][CLI] Improve overall startup time