--- a/vllm/entrypoints/cli/openai.py
+++ b/vllm/entrypoints/cli/openai.py
@@ -2,20 +2,18 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 # Commands that act as an interactive OpenAI API client
 
-from __future__ import annotations
-
 import argparse
 import os
 import signal
 import sys
-import typing
+from typing import TYPE_CHECKING, Optional
 
 from openai import OpenAI
 from openai.types.chat import ChatCompletionMessageParam
 
 from vllm.entrypoints.cli.types import CLISubcommand
 
-if typing.TYPE_CHECKING:
+if TYPE_CHECKING:
     from vllm.utils import FlexibleArgumentParser
 
 
@@ -46,7 +44,8 @@ def _interactive_cli(args: argparse.Namespace) -> tuple[str, OpenAI]:
     return model_name, openai_client
 
 
-def chat(system_prompt: str | None, model_name: str, client: OpenAI) -> None:
+def chat(system_prompt: Optional[str], model_name: str,
+         client: OpenAI) -> None:
     conversation: list[ChatCompletionMessageParam] = []
     if system_prompt is not None:
         conversation.append({"role": "system", "content": system_prompt})