--- a/vllm/entrypoints/cli/run_batch.py
+++ b/vllm/entrypoints/cli/run_batch.py
@@ -6,14 +6,15 @@ from __future__ import annotations
 import argparse
 import asyncio
 import importlib.metadata
+import typing
 
 from vllm.entrypoints.cli.types import CLISubcommand
-from vllm.entrypoints.openai.run_batch import main as run_batch_main
-from vllm.entrypoints.openai.run_batch import make_arg_parser
 from vllm.entrypoints.utils import (VLLM_SUBCMD_PARSER_EPILOG,
                                     show_filtered_argument_or_group_from_help)
 from vllm.logger import init_logger
-from vllm.utils import FlexibleArgumentParser
+
+if typing.TYPE_CHECKING:
+    from vllm.utils import FlexibleArgumentParser
 
 logger = init_logger(__name__)
 
@@ -24,6 +25,8 @@ class RunBatchSubcommand(CLISubcommand):
 
     @staticmethod
     def cmd(args: argparse.Namespace) -> None:
+        from vllm.entrypoints.openai.run_batch import main as run_batch_main
+
         logger.info("vLLM batch processing API version %s",
                     importlib.metadata.version("vllm"))
         logger.info("args: %s", args)
@@ -44,6 +47,8 @@ class RunBatchSubcommand(CLISubcommand):
     def subparser_init(
             self,
             subparsers: argparse._SubParsersAction) -> FlexibleArgumentParser:
+        from vllm.entrypoints.openai.run_batch import make_arg_parser
+
         run_batch_parser = subparsers.add_parser(
             "run-batch",
             help="Run batch prompts and write results to file.",