--- a/vllm/triton_utils/importing.py
+++ b/vllm/triton_utils/importing.py
@@ -74,10 +74,6 @@ class TritonPlaceholder(types.ModuleType):
         self.heuristics = self._dummy_decorator("heuristics")
         self.Config = self._dummy_decorator("Config")
         self.language = TritonLanguagePlaceholder()
-        logger.warning_once(
-            "Triton is not installed. Using dummy decorators. "
-            "Install it via `pip install triton` to enable kernel"
-            " compilation.")
 
     def _dummy_decorator(self, name):