This PR performs lazy loading for importing vllm, where it tries to cleanup a lot of eager heavy imports to help improve general CLI startup experience

This PR does:
- Lazy load imports `vllm/__init__.py` to avoid unwanted side-effect. I have checked that most of these imports will be run during serve/engine initialization
- Move a lot of the configuration validator into `@model_validator` hooks
- Cached the compute kwargs in `arg_utils.py`, given that these are currently being called multiples times (we only need this for CLI, so most of the time, it is safe to have this cache. I used a lru_cache size of 30 preemptively to make sure we have enough room for future configuration if needed)
- Lazy load all CLI imports, given that we don't need to setup eagerly during startup.
- remove triton_utils side effect in utils, given that this doesn't seem to do anything (for all of the vllm.triton_utils import, it should validate HAS_TRITON accordingly).

With this PR, it reduces the vllm CLI by around 2 seconds which we still have a lot of room for improvement (a part of #19824 to improve general startup UX)

Here is the hyperfine for `vllm -h`. Python 3.11, 8xH100 instances on LambdaLabs:

with this PR

```bash
➜ hyperfine 'vllm -h' --warmup 3 --runs 100
Benchmark 1: vllm -h
  Time (mean ± σ):      6.705 s ±  0.016 s    [User: 14.560 s, System: 0.571 s]
  Range (min … max):    6.676 s …  6.798 s    100 runs

➜ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      1.344 s ±  0.004 s    [User: 9.608 s, System: 0.181 s]
  Range (min … max):    1.336 s …  1.356 s    100 runs
```

On current [main](https://github.com/vllm-project/vllm/commit/4c409cabc2c1c432ba670029990bd59e6bbf1479)

```bash
➜ hyperfine 'vllm -h' --warmup 3 --runs 100
Benchmark 1: vllm -h
  Time (mean ± σ):      8.635 s ±  0.020 s    [User: 16.423 s, System: 0.637 s]
  Range (min … max):    8.544 s …  8.730 s    100 runs

➜ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      5.194 s ±  0.060 s    [User: 13.067 s, System: 0.552 s]
  Range (min … max):    5.139 s …  5.568 s    100 runs
```

Signed-off-by: Aaron Pham <contact@aarnphm.xyz>